fix: strip repo to bare minimum like Smithery cookbook
Removed ALL extra files to match cookbook exactly:
- Deleted .well-known/mcp.json, icon.svg, LICENSE
- Deleted generators.py, models.py, parser.py, templates.py
- Renamed server_minimal.py to server.py
- Simplified __init__.py to just a comment

This is an exact copy of the cookbook structure to debug 502.
