+++ src/devplan_mcp/__init__.py
@@ -1,2 +1,1 @@
-# DevPlan MCP Server - Generate development plans via Model Context Protocol
-__version__ = "0.1.0"
+# DevPlan MCP Server Package
